Support for timestamp change script numbering
Made modifications to allow use of timestamp numbering instead of sequential numbering (sequential numbering can still be used).  Modifications include:
* Converted change script number from int to long
* Created an ant task to automatically create a new timestamped change script
* Created a command line interface that can be run to create a new timestamped change script.

The timestamp is currently not configurable and in the format yyyyMMddHHmm.

+++ dbdeploy-core/src/main/java/com/dbdeploy/AppliedChangesProvider.java
@@ -3,5 +3,5 @@ package com.dbdeploy;
 import java.util.List;
 
 public interface AppliedChangesProvider {
-	List<Integer> getAppliedChanges();
+	List<Long> getAppliedChanges();
 }
