+++ dbdeploy-dist/src/main/scripts/createSchemaVersionTable.ora.sql
@@ -1,5 +1,5 @@
 CREATE TABLE changelog (
-  change_number INTEGER NOT NULL,
+  change_number NUMBER(22,0) NOT NULL,
   complete_dt TIMESTAMP NOT NULL,
   applied_by VARCHAR2(100) NOT NULL,
   description VARCHAR2(500) NOT NULL
